added policy to post

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -31,6 +32,8 @@ class PostController extends Controller
 
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('update', $post);
+        // validate
         $data = $request->validate([
             'title' => 'required|string|max:255',
             'body' => 'required',
@@ -50,6 +53,9 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        // authorize
+        Gate::authorize('delete',$post);
+
         $post->delete();
         return response()->json(null, 204);
     }
